Split SICP layout and page
